requirements.txt
Flask>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
Flask-Login>=0.6.3
Flask-Dance>=7.0.0
blinker>=1.4.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0